refactor readme testers
