+++ src/testUtil.js
@@ -1,0 +1,96 @@
+// Copyright 2023 Cisco Systems, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+// Copyright 2023 Cisco Systems, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+import fs from 'fs';
+import CliCore from "./CliCore.js";
+
+/**
+ * Regular expression for command extraction from README.md file. This program finds all the markup code blocks
+ * that begin and end with ``` (markdown syntax for code block). It first checks for an optional note
+ * (starting with '> .note') before extracting the cli command and the response. It then runs the cli
+ * command and compares the response to what is in the README.md. This ensures that the README is always accurate.
+ *
+ * /^(?:> \.note (?<note>.+)[\r\n])?^> \.(?<command>.+[\r\n])((?<expectedResponse>(?:(?!^>|```)[\s\S])*))$/gm
+ *
+ * Breakdown:
+ *
+ * ^: Start of a line (because we're using the 'm' flag which makes ^ and $ match start and end of lines respectively)
+ *
+ * (?:...): Non-capturing group.
+ *
+ * > : Matches the ">" character literally. This assumes that your command lines in the README start with ">".
+ *
+ * \.note : Matches the ".note" literal string, denoting a note for the command.
+ *
+ * (?<note>.+[\r\n]): Optional named capturing group 'note'. Matches any number of characters (using .+), followed by a line break (using [\r\n]).
+ *
+ * \. : Matches the "." character literally.
+ *
+ * (?<command>.+[\r\n]): Named capturing group 'command'. Matches any number of characters (using .+), followed by a line break (using [\r\n]).
+ *
+ * ((?<expectedResponse>(?:(?!^>|```)[\s\S])*)): Named capturing group 'expectedResponse'. Matches any number of characters that do not start a new command line or a markdown code block.
+ *
+ * Inside 'expectedResponse':
+ *    (?:...): Non-capturing group, used to group these elements together without remembering their matched content.
+ *    (?!^>|```): Negative lookahead. Asserts that what follows is not the start of a new command line (^>) or a markdown code block (```).
+ *    [\s\S]: Matches any character, including newlines. \s matches any whitespace character, and \S matches any non-whitespace character. Together, they match any character.
+ *    *: Matches zero or more of the preceding element.
+ *
+ * $: End of a line (because we're using the 'm' flag which makes ^ and $ match start and end of lines respectively)
+ *
+ * g: Global flag, to find all matches in the string, not just the first one.
+ *
+ * m: Multiline flag, to allow ^ and $ to match the start and end of lines, not just the start and end of the whole string.
+ */
+const markdownContent = fs.readFileSync("README.md", 'utf-8');
+const commandRegex = /^(?:> \.note (?<note>.+)[\r\n])?^> \.(?<command>.+[\r\n])((?<expectedResponse>(?:(?!^>|```)[\s\S])*))$/gm;
+
+export function getTestDataFromReadme() {
+    let match;
+    const cliCore = new CliCore();
+    const testData = [];
+
+    while ((match = commandRegex.exec(markdownContent)) !== null) {
+        const command = match.groups.command.trim();
+        const expectedResponseString = match.groups.expectedResponse.trim();
+        const note = match.groups.note?.trim();
+        const args = command.split(' ');
+
+        if (args.length > 0) {
+            const cmdName = args.shift();
+            const method = cliCore[cmdName];
+
+            if (typeof method === 'function') {
+                testData.push({ cmdName, args, expectedResponseString, note });
+            } else {
+                throw Error(`Unknown command: .${cmdName}`);
+            }
+        }
+    }
+
+    return {cliCore, testData};
+}
